# backoffice/docs/04-openapi.yaml
openapi: 3.0.3
info:
  title: Planet Ultra Resident Management API
  version: 1.0.0
  description: |
    Production API baseline for mobile app + backoffice SaaS.
servers:
  - url: https://api.planetultra.app/v1
...
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      tags: [Users]
      summary: List users
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - name: role
          in: query
          schema:
            $ref: '#/components/schemas/UserRole'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/UserStatus'
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsers'
    post:
      tags: [Users]
      summary: Create user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateInput'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /users/{id}:
    get:
      tags: [Users]
      summary: Get user detail
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: User detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    patch:
      tags: [Users]
      summary: Update user
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateInput'
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
  /apartments:
    get:
      tags: [Apartments]
      summary: List apartments
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Search'
        - name: occupancyStatus
          in: query
          schema:
            $ref: '#/components/schemas/ApartmentOccupancyStatus'
      responses:
        '200':
          description: Apartment list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedApartments'
    post:
      tags: [Apartments]
      summary: Create apartment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApartmentCreateInput'
      responses:
        '201':
          description: Apartment created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Search:
      name: search
      in: query
      schema:
        type: string
  schemas:
...
    UserRole:
      type: string
      enum: [Admin, Manager, Security, Resident]
    UserStatus:
      type: string
      enum: [Invited, Active, Suspended, Deactivated]
    ApartmentOccupancyStatus:
      type: string
      enum: [Occupied, Vacant, Maintenance, Inactive]
...
    User:
      type: object
      required: [id, fullName, email, role, status, createdAt, updatedAt]
      properties:
        id: { type: string }
        fullName: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        role: { $ref: '#/components/schemas/UserRole' }
        status: { $ref: '#/components/schemas/UserStatus' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    UserCreateInput:
      type: object
      required: [fullName, email, role]
      properties:
        fullName: { type: string }
        email: { type: string, format: email }
        phone: { type: string }
        role: { $ref: '#/components/schemas/UserRole' }
        status: { $ref: '#/components/schemas/UserStatus' }
    UserUpdateInput:
      type: object
      properties:
        fullName: { type: string }
        phone: { type: string }
        status: { $ref: '#/components/schemas/UserStatus' }
    Apartment:
      type: object
      required: [id, tower, unitNumber, floor, occupancyStatus, createdAt, updatedAt]
      properties:
        id: { type: string }
        tower: { type: string }
        unitNumber: { type: string }
        floor: { type: integer }
        occupancyStatus: { $ref: '#/components/schemas/ApartmentOccupancyStatus' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ApartmentCreateInput:
      type: object
      required: [tower, unitNumber, floor]
      properties:
        tower: { type: string }
        unitNumber: { type: string }
        floor: { type: integer }
...
    PaginatedUsers:
      type: object
      required: [data, page, limit, total]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
    PaginatedApartments:
      type: object
      required: [data, page, limit, total]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Apartment'
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }